Fixed typo

+++ Sources/Netlink/AttributeDecoder.swift
@@ -867,7 +867,7 @@ extension Int64: NetlinkAttributeDecodable {
     
     public init?(attributeData data: Data) {
         
-        guard data.count == MemoryLayout<Int32>.size
+        guard data.count == MemoryLayout<Int64>.size
             else { return nil }
         
         self.init(bytes: (data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7]))
